requirements.txt
selenium
webdriver-manager
google
google-genai
pyaudio
openai-whisper
